Type of amount changed to double

+++ Abschlussprojekt/src/main/java/Bendispository/Abschlussprojekt/service/TransactionService.java
@@ -41,7 +41,7 @@ public class TransactionService {
 
 
     public boolean lenderApproved(Request request){
-        int deposit = request.getRequestedItem().getDeposit();
+        double deposit = request.getRequestedItem().getDeposit();
         Person requester = request.getRequester();
         if(proPaySubscriber.checkDeposit(deposit,
                                          requester.getUsername())) {
@@ -105,7 +105,7 @@ public class TransactionService {
 
         //zurückbuchung deposit
 
-        int amount = leaseTransaction.getDuration() * leaseTransaction.getItem().getCostPerDay();
+        double amount = leaseTransaction.getDuration() * leaseTransaction.getItem().getCostPerDay();
         makePayment(leaser, lender, amount, leaseTransaction, PaymentType.RENTPRICE);
 
         // 1. zeitgemäß?
@@ -122,13 +122,13 @@ public class TransactionService {
             leaseTransaction.setTimeframeViolation(true);
             leaseTransaction.setLengthOfTimeframeViolation(timeViolation);
 
-            int amount = leaseTransaction.getItem().getCostPerDay() * timeViolation;
+            double amount = leaseTransaction.getItem().getCostPerDay() * timeViolation;
             makePayment(leaser, lender, amount, leaseTransaction, PaymentType.DAMAGES);
 
         }
     }
 
-    private void makePayment(Person leaser, Person lender, int amount, LeaseTransaction leaseTransaction, PaymentType type){
+    private void makePayment(Person leaser, Person lender, double amount, LeaseTransaction leaseTransaction, PaymentType type){
         PaymentTransaction paymentTransaction = new PaymentTransaction(leaser, lender, amount);
         paymentTransaction.setType(type);
         paymentTransaction.setLeaseTransaction(leaseTransaction);
